Add JwtBearer Authentication in Startup.cs
Added Microsoft.IdentityModel.Tokens namespace and JwtBearer authentication to Startup.cs. The purpose of this change was to add additional layer of security by validating JWT tokens. These changes will enable the server to authenticate and authorize users based on JWT tokens, making our app more secure and trustworthy. As per modifications, now the server will not validate the lifetime of JWT tokens. This is to allow the tokens to be persistently stored on client's machine until they log out manually.