Feat: Add picture header to tournament cards
Images have been added to the tournament card headers in 'TournamentMainTab.razor' and 'AllTournaments.razor'. The layout of the header has also changed: the header now stacks the image and title vertically, and centers them, with space added between the image and title. This is to enhance the visual appeal of the cards. Furthermore, the authorization view in 'Tournaments.razor' has been temporarily disabled. Finally, an empty line was added to 'Program.cs' which has no effect on the code execution or structure.