Refact: "Exclude current user from player search results in tournament"
In the TournamentController, the SearchToInvitePlayer method has been modified to exclude the current user from the search results. The change was made to ensure that a user cannot invite themselves to the tournament. The userId of the currently logged in user is retrieved and checked against the UserId of each player during the search query.