Prevent player removal from tournament matches and improve match creation
In MatchController, added a condition to deny requests for player removal from tournament matches as it would disrupt the structure of the tournament. In GenerateRoundRobinPhaseCommandHandler, switched to RoundRobinMatchGenerator for creating matches to enhance code readability and maintainability. Also, marked the match players as confirmed during match creation to reflect their committed participation.