Feat: Add Elo average to tournament display
Added a new feature to display the average Elo rating of the participants in each tournament. This allows players to get a sense of the competition level in the tournament. The Elo rating was added to the tournament response model, calculated in the tournament controller, and displayed in the AllTournaments.razor component.