Refact: Add ImageUri property to Tournament model
The ImageUri property has been added to the Tournament model, response, and controllers which allow users to provide or retrieve a specific image URL for a tournament. This was done to increase flexibility in tournament image sources. Previously, the image source was hardcoded, now it is dynamic based on the provided Image URL. This change affects the Shared Responses, Client Components, Server Controllers, Server Models, and Server Data Migrations.