FEat: Update match start time calculation in round robin generator
The match start time is now calculated with the 'AddHours' method instead of the 'AddMinutes' method. The calculation now takes the round number into account and allocates more time (0.50 hours) to each round. This change was necessary to more accurately reflect the duration of matches and avoid overlaps.

Additionally, array rotation logic and a few syntax adjustments have been made for readability.