Feat: Add tournament details to MatchDetail page
Added tournament details to the MatchDetail page and updated the MatchResponse model to include more tournament information. This will make it easier for users to see important tournament information such as the tournament name, phase, round, and associated image directly on the MatchDetail page. This is achieved by a new div section on the MatchDetail view that displays these details. If a match is linked with a tournament, users can also navigate to the tournament's details page. The Server Handlers have been modified to set and get the appropriate fields.