Reduce Elo point loss on match defeat
Adjusted the Elo rating system's penalty for losing a match. Previously, the loss of Elo points was quite harsh, effectively doubling if the team lost. However, this was found to be too punitive and discouraged competitive play. With this change, the Elo loss of the losing team has been reduced from doubling to being multiplied by a factor of 1.8. This should foster a healthier competitive environment by lessening the potential negative impact of a single match loss on a player's ranking.