"Update Match and Tournament controls for better user experience."
This change updates the controllers in both Match and Tournament to improve the user experience. In the MatchManagerController.cs, the method 'SingleOrDefault' which throws an exception if there is more than one element in the sequence is replaced with 'FirstOrDefault'. This will instead return null and will prevent potential errors from crashing the application if there are more players with the same ID.

In the TournamentMatch.razor component, an extra condition is added to prevent deleting results of finished matches which will prevent inconsistent game states. Also, added helpful UI feedback to guide users through the process of confirming match results in round robin phase.

In TournamentController.cs, added a new endpoint which allows users to confirm results for all matches in a round-robin phase at once, this reduce repetitive actions and thus enhancing user experience. Lastly, 'AsNoTracking' was added during data fetch from the database for performance improvements as the tracked entities are not needed in these scenerios.