Refactor round robin phase loading logic
This commit changes the round robin phase loading strategy in Tournaments.

Now instead of loading all the round robin matches in the TournamentController's Get method,
they are retrieved in a separate specific API call
which is made within the TournamentCompetition's OnInitializedAsync lifecycle method.
This changes help in reducing the loading time as API call is only made when necessary
and not every time the TournamentController's Get method is called.
It also improves the encapsulation and responsibility segregation of the Tournament main component and its child TournamentCompetition component.

The API call is also now made with respect to the user's authentication state,
using either an anonymous or authenticated API client as required.