Add finals loading and error handling in TournamentFinals component
This commit introduces changes in several files related to the Tournament Finals component. The primary change is in the TournamentFinals.razor where a loading state has been added to handle the fetching of the final matches. If there are no matches yet, a meaningful message is displayed. Error handling has also been improved, with error messages now being shown to the user in the form of a toast notification.

Also, new fields and methods related to the finals phase of the tournament have been implemented in several components and response classes, like TournamentAdmin.razor, TournamentCompetition.razor and TournamentController.cs. A new response class `TournamentFinalsMatchResponse` is introduced to facilitate the communication of the finals phase data.

This modification enhances user experience by providing a loading state and better error handling during the tournament, specifically in the finalists' presentation. Additionally, it facilitates front-end and back-end communication regarding the finals phase data.