Add RoundRobinType to tournament details and requests
Added the RoundRobinType enumeration to both TournamentDetailResponse and TournamentRequest. This change allows users to specify the type of round-robin tournament when either creating a new tournament or editing an existing one. This is done via a new select field in both the CreateTournamentModal and EditTournamentModal. The options are 'Random groups' and 'Best players lead groups'.