Add 'RoundRobinType' to tournament configuration
This commit introduces the 'RoundRobinType' attribute to the tournament configuration in 'EditTournamentModal.razor'. The server-side logic for dividing teams into groups in 'GenerateRoundRobinPhaseCommandHandler.cs' has been updated to account for the new attribute. 'RoundRobinType.BestPlayersLeadGroups' algorithm is implemented to sort teams based on Elo. This change allows tournament organizers to choose how the teams will be divided in the round-robin phase of the tournament. It increases flexibility in tournament setup, especially useful when team strength (Elo) varies significantly.