Add functionality to identify match winner
Updated the code in the 'MatchController.cs' and 'CalculateEloResultAfterMatchCommandHandler.cs' files to add the ability to identify and calculate the winner of a match based off the total score of each team. This modifies the way team winning is determined and the calculation for the new Elo of each player after a match. The winner of a match is obtained by comparing the scores of both teams, and then updating the teams' respective scores in the database. The calculation of a player's Elo score after a match has been also refined, basically rounding it to the nearest integer.