Remove in-progress alert from TournamentCompetition
The in-progress and finished alerts have been removed from the TournamentCompetition component. This change was implemented as it was causing confusion for users who weren't sure if a phase was ongoing or concluded based on the alerts displayed. Alerts will be replaced with a more intuitive system in subsequent updates.