Add email field to the new player creation
In the registration and external login forms, realized that the new Player object created did not have the Email field populated. This might cause issues later when we need to use email for player-related operations such as password reset or notifications. Therefore, added the Email field to the new Player object creation process.