Update maintenance message with migration info
The maintenance message in App.razor has been updated to include detailed information about an application migration. The message now informs users that the application has moved to a new address, provides a link to the new location, and instructs them on how to transfer their data and use the new platform. In addition, it also includes updates on future developments and improvements.